requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
uuid==1.30
Werkzeug==2.0.1
